discord-huddle/requirements.txt
requests>=2.31
python-dotenv>=1.0
discord.py>=2.3
